link_bio/requirements.txt
reflex==0.3.10
python-dotenv
supabase
configcat-client